chore(coordinate.py): move module into `geometry/` subdirectory
