+++ puffkit/__init__.py
@@ -7,9 +7,9 @@ print(f"puffkit {__version__}")
 from puffkit.app import PkApp
 from puffkit.color import palettes
 from puffkit.color.color import ColorValue, PkColor
-from puffkit.coordinate import PkCoordinate
+from puffkit.geometry.coordinate import PkCoordinate
+from puffkit.geometry.rect import PkRect
 from puffkit.object import PkObject
-from puffkit.rect import PkRect
 from puffkit.scene import PkScene
 from puffkit.subsurface import PkSubSurface
 from puffkit.surface import PkSurface
